add wrapper for imgui PlotLines widget
This patch only wraps the PlotLines widget from imgui, but the histogram
and other plotting widgets could be wrapped also.

This also adds an example in examples/plot showing both g.PlotLines()
and g.PlotLinesV().

Future work: the imgui plotting widget is a little limited, it would be
nice for GIU to support a more robust plotting system. This could
perhaps be done from scratch based on a Canvas widget, possibly by
writing a back-end for gonum/plot.