widgets/ImageWithRGBAWidget: store texture in widget struct
when image with rgba is created, the texture (giu.Texture) is created and stored in Context. But when image isn't built, the context is disposed and texture is lost. while presenting a few images (and showing them dependnig on e.g. on slider/combo) a delay while loading images is visible
the solution is to store imageWidget in a variable