mainthread: disable on windows
in reference to https://github.com/AllenDang/giu/issues/735#issuecomment-1951419027 we have no mainthread
library that works on windows. Due to this fact I think that the best solution is to disable mainthread completely for winows until we find some working solution.

I'll create an issue about this as soon as this change is merged to master